pyproject.toml
[project]
name = "FiveNightsAtWonderland"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.3",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]
